+++ src/tiptap-editor.vue
@@ -104,7 +104,7 @@
 
       <v-button
         v-if="editorExtensions.includes('code')"
-        v-tooltip="t('wysiwyg_options.codeblock') + ' - ' + translateShortcut(['meta', 'e'])"
+        v-tooltip="t('code') + ' - ' + translateShortcut(['meta', 'e'])"
         small
         icon
         :disabled="props.disabled"
@@ -140,7 +140,13 @@
             :active="editor.isActive('heading')"
             @click="toggle"
           >
-            <icon-heading />
+            <icon-h1 v-if="editor.isActive('heading', { level: 1 })" />
+            <icon-h2 v-if="editor.isActive('heading', { level: 2 })" />
+            <icon-h3 v-if="editor.isActive('heading', { level: 3 })" />
+            <icon-h4 v-if="editor.isActive('heading', { level: 4 })" />
+            <icon-h5 v-if="editor.isActive('heading', { level: 5 })" />
+            <icon-h6 v-if="editor.isActive('heading', { level: 6 })" />
+            <icon-heading v-if="!editor.isActive('heading')" />
           </v-button>
         </template>
         <v-list>
@@ -205,7 +211,55 @@
         <icon-double-quotes-r />
       </v-button>
 
+      <v-button
+        v-if="editorExtensions.includes('codeBlock')"
+        v-tooltip="t('wysiwyg_options.codeblock') + ' - ' + translateShortcut(['meta', 'alt', 'c'])"
+        small
+        icon
+        :disabled="props.disabled"
+        :active="editor.isActive('codeBlock')"
+        @click="editor.chain().focus().toggleCodeBlock().run()"
+      >
+        <icon-code-box-line />
+      </v-button>
+
       <div class="divider" />
+
+      <v-menu v-if="editorExtensions.includes('textAlign')" show-arrow placement="bottom-start">
+        <template #activator="{ toggle }">
+          <v-button v-tooltip="t('tiptap.text_align')" :disabled="props.disabled" small icon active @click="toggle">
+            <template v-for="opt in alignOptions" :key="opt.align">
+              <component v-if="editor.isActive({ textAlign: opt.align })" :is="opt.icon" />
+            </template>
+          </v-button>
+        </template>
+        <v-list>
+          <v-list-item
+            v-for="opt in alignOptions"
+            :key="opt.align"
+            clickable
+            :active="editor.isActive({ textAlign: opt.align })"
+            @click="
+              editor.isActive({ textAlign: opt.align })
+                ? editor.chain().focus().unsetTextAlign().run()
+                : editor.chain().focus().setTextAlign(opt.align).run()
+            "
+          >
+            <v-list-item-icon>
+              <component :is="opt.icon" />
+            </v-list-item-icon>
+            <v-list-item-content>
+              <v-text-overflow :text="opt.text" />
+            </v-list-item-content>
+            <v-list-item-hint>{{ opt.shortcut }}</v-list-item-hint>
+          </v-list-item>
+          <v-list-item clickable @click="editor.chain().focus().unsetTextAlign().run()">
+            <v-list-item-content>
+              <v-text-overflow :text="t('wysiwyg_options.alignnone')" />
+            </v-list-item-content>
+          </v-list-item>
+        </v-list>
+      </v-menu>
 
       <v-button
         v-if="editorExtensions.includes('horizontalRule')"
@@ -216,6 +270,29 @@
         @click="editor.chain().focus().setHorizontalRule().run()"
       >
         <icon-separator />
+      </v-button>
+
+      <v-button
+        v-if="editorExtensions.includes('link')"
+        v-tooltip="t('wysiwyg_options.link') + ' - ' + translateShortcut(['meta', 'k'])"
+        small
+        icon
+        :disabled="props.disabled"
+        :active="editor.isActive('link')"
+        @click="linkOpen"
+      >
+        <icon-link />
+      </v-button>
+
+      <v-button
+        v-if="editorExtensions.includes('link')"
+        v-tooltip="t('wysiwyg_options.unlink')"
+        small
+        icon
+        :disabled="props.disabled || !editor.isActive('link')"
+        @click="editor.chain().focus().extendMarkRange('link').unsetLink().run()"
+      >
+        <icon-unlink />
       </v-button>
 
       <v-button
@@ -268,10 +345,38 @@
 
     <editor-content class="tiptap-editor__content" :editor="editor" />
 
-    <div class="tiptap-editor__info">
+    <v-dialog v-model="linkDrawerOpen">
+      <v-card class="card">
+        <v-card-title class="card-title">{{ t("wysiwyg_options.link") }}</v-card-title>
+        <v-card-text>
+          <div class="grid">
+            <div class="field">
+              <div class="type-label">{{ t("url") }}</div>
+              <v-input v-model="linkHref" :placeholder="t('url_placeholder')"></v-input>
+            </div>
+            <div class="field">
+              <div class="type-label">{{ t("open_link_in") }}</div>
+              <v-input v-model="linkTarget"></v-input>
+            </div>
+          </div>
+        </v-card-text>
+        <v-card-actions>
+          <v-button secondary @click="linkClose">{{ t("cancel") }}</v-button>
+          <v-button :disabled="linkHref === null" @click="linkSave">{{ t("save") }}</v-button>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+
+    <div class="tiptap-editor__info" v-if="editorExtensions.includes('characterCount')">
       <div v-if="editorExtensions.includes('characterCount')">
-        {{ t("tiptap.count_chars", editor.storage.characterCount.characters()) }},
-        {{ t("tiptap.count_words", editor.storage.characterCount.words()) }}
+        <template v-if="!editor.storage.characterCount.characters() && !editor.storage.characterCount.words()">
+          ∅
+        </template>
+        <template v-else>
+          {{ t("tiptap.count_words", editor.storage.characterCount.words()) }},
+          {{ t("tiptap.count_chars", editor.storage.characterCount.characters()) }}
+          <template v-if="props.characterCountLimit"> / {{ props.characterCountLimit }}</template>
+        </template>
       </div>
     </div>
   </div>
@@ -377,11 +482,14 @@
   }
 
   &__content {
-    min-height: 230px;
     font-family: var(--family-sans-serif);
     font-weight: 400;
-    margin: var(--input-padding) 0;
-    padding: 0 var(--input-padding);
+
+    .ProseMirror {
+      min-height: 230px;
+      margin: var(--input-padding) 0;
+      padding: 0 var(--input-padding);
+    }
 
     * {
       margin: revert;
@@ -390,9 +498,32 @@
       line-height: revert;
     }
 
+    a {
+      color: var(--blue);
+      text-decoration: underline;
+    }
+
     b,
     strong {
       font-weight: 700;
+    }
+
+    pre,
+    code {
+      background: var(--background-inverted);
+      color: var(--foreground-inverted);
+      border-radius: 0.3em;
+    }
+
+    pre {
+      padding: 0.4em 0.6em;
+    }
+    code {
+      padding: 0 0.2em;
+    }
+
+    pre > code {
+      padding: 0;
     }
 
     .ProseMirror p.is-editor-empty:first-child::before {
@@ -408,12 +539,11 @@
 
 <script setup lang="ts">
 import { Editor, EditorContent, BubbleMenu } from "@tiptap/vue-3";
-import { Placeholder } from "@tiptap/extension-placeholder";
 import { onBeforeUnmount, watch } from "vue";
 import { useI18n } from "vue-i18n";
 import { translateShortcut } from "./utils/translate-shortcut";
 import type { TypeType, ValueType } from "./types";
-import { extensions } from "./extensions";
+import { loadExtensions } from "./extensions";
 import messages from "./messages.json";
 import IconArrowGoBackLine from "./icons/arrow-go-back-line.vue";
 import IconParagraph from "./icons/paragraph.vue";
@@ -424,6 +554,12 @@ import IconItalic from "./icons/italic.vue";
 import IconStrikethrough from "./icons/strikethrough.vue";
 import IconUnderline from "./icons/underline.vue";
 import IconHeading from "./icons/heading.vue";
+import IconH1 from "./icons/h1.vue";
+import IconH2 from "./icons/h2.vue";
+import IconH3 from "./icons/h3.vue";
+import IconH4 from "./icons/h4.vue";
+import IconH5 from "./icons/h5.vue";
+import IconH6 from "./icons/h6.vue";
 import IconSeparator from "./icons/separator.vue";
 import IconArrowGoForwardLine from "./icons/arrow-go-forward-line.vue";
 import IconCodeLine from "./icons/code-line.vue";
@@ -433,34 +569,86 @@ import IconFormatClear from "./icons/format-clear.vue";
 import IconMarkPenLine from "./icons/mark-pen-line.vue";
 import IconSubscript from "./icons/subscript.vue";
 import IconSuperscript from "./icons/superscript.vue";
+import IconCodeBoxLine from "./icons/code-box-line.vue";
+import IconAlignLeft from "./icons/align-left.vue";
+import IconAlignCenter from "./icons/align-center.vue";
+import IconAlignRight from "./icons/align-right.vue";
+import IconAlignJustify from "./icons/align-justify.vue";
+import IconLink from "./icons/link.vue";
+import IconUnlink from "./icons/unlink.vue";
+import type { CharacterCountOptions } from "@tiptap/extension-character-count";
+import type { TextAlignOptions } from "@tiptap/extension-text-align";
+import textAlign from "./extensions/text-align";
+import characterCount from "./extensions/character-count";
+import type { PlaceholderOptions } from "@tiptap/extension-placeholder";
+import placeholder from "./extensions/placeholder";
+import { useLink } from "./composables/link";
 
 const { t } = useI18n({ messages });
 
-const props = withDefaults(
-  defineProps<{
-    value: ValueType | null;
-    type: TypeType;
-    disabled: boolean;
-    placeholder: string | null;
-  }>(),
-  { value: null, disabled: false, placeholder: null }
-);
+interface Props {
+  value: ValueType | null;
+  type: TypeType;
+  disabled: boolean;
+  autofocus: boolean;
+  extensions: string[] | null;
+  // extensions options
+  placeholder: PlaceholderOptions["placeholder"];
+  textAlignTypes: TextAlignOptions["types"];
+  characterCountLimit: CharacterCountOptions["limit"];
+  characterCountMode: CharacterCountOptions["mode"];
+}
+
+const props = withDefaults(defineProps<Props>(), {
+  value: null,
+  disabled: false,
+  autofocus: false,
+  extensions: null,
+  placeholder: () => placeholder.defaults.placeholder,
+  textAlignTypes: () => textAlign.defaults.types,
+  characterCountLimit: () => characterCount.defaults.limit,
+  characterCountMode: () => characterCount.defaults.mode,
+});
 
 const emit = defineEmits<{
   (e: "input", value: ValueType): void;
 }>();
 
-extensions.push(
-  Placeholder.configure({
-    placeholder: props.placeholder,
-  })
-);
+const alignOptions = [
+  {
+    align: "left",
+    icon: IconAlignLeft,
+    text: t("wysiwyg_options.alignleft"),
+    shortcut: translateShortcut(["meta", "shift", "l"]),
+  },
+  {
+    align: "center",
+    icon: IconAlignCenter,
+    text: t("wysiwyg_options.aligncenter"),
+    shortcut: translateShortcut(["meta", "shift", "e"]),
+  },
+  {
+    align: "right",
+    icon: IconAlignRight,
+    text: t("wysiwyg_options.alignright"),
+    shortcut: translateShortcut(["meta", "shift", "r"]),
+  },
+  {
+    align: "justify",
+    icon: IconAlignJustify,
+    text: t("wysiwyg_options.alignjustify"),
+    shortcut: translateShortcut(["meta", "shift", "j"]),
+  },
+];
+
+const extensions = loadExtensions(props);
 
 const editor = new Editor({
   editable: !props.disabled,
   content: props.value,
   extensions,
-  onUpdate: () => {
+  autofocus: props.autofocus,
+  onUpdate: ({ editor }) => {
     switch (props.type) {
       case "json":
         emit("input", editor.getJSON());
@@ -473,6 +661,8 @@ const editor = new Editor({
 });
 
 const editorExtensions = editor.extensionManager.extensions.map((ext) => ext.name);
+
+const { linkDrawerOpen, linkHref, linkTarget, linkOpen, linkClose, linkSave } = useLink(editor);
 
 watch(
   () => props.value,